requirements.txt
# Create requirements.txt with all needed packages
cat > requirements.txt << EOL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
nltk>=3.8.0
datasets>=2.12.0
tqdm>=4.65.0
pyyaml>=6.0
jupyter>=1.0.0
pytest>=7.3.0
scikit-learn>=1.2.0
astor>=0.8.1
transformers>=4.28.0
einops>=0.6.1
EOL

# Then install
